refactor(storage): hoist engine-capability descriptor cluster to proximadb-storage-ports (root-decomp Slice D)
Move the cohesive ~600-line, ~10-type pure-metadata engine-capability
descriptor cluster out of the root crate into the foundation
proximadb-storage-ports crate, clearing the last root-dep
(CapabilityFactory) of src/storage/traits/mod.rs so the engine-port-traits
module can move to its own crate (unblocks the 188K-symbol engines
extraction).

Cluster members (moved as a unit to ONE destination crate):
- StorageEngineStrategy enum + StorageFormatStrategy alias (from traits/types.rs)
- ScanCapabilities struct (from scan_strategy.rs)
- FlushThresholds + CompactionHeuristics + their Default impls
- EngineCapabilities trait
- 7 per-engine capability structs + impls (Sst/Helix/Viper/Swift/Nova/Tst/Raptor)
- EngineBundle<E> generic + impl
- CapabilityFactory + impl (incl. from_proto_engine bridge)
- the cluster's #[cfg(test)] unit tests

Verified pure-metadata (no engine-type coupling): grep for
SstEngine|ViperEngine|HelixEngine|NovaEngine|SwiftEngine|RaptorEngine|TstEngine
across the three cluster files returns empty. The cluster's only deps are
each other + proto (CompressionAlgorithm, StorageEngine) + serde + std.

Old import paths preserved via pub-use re-export shims so every existing
caller resolves unchanged:
- src/storage/trait_components/capabilities.rs -> re-exports all moved types
- src/storage/scan_strategy.rs -> re-exports ScanCapabilities
- src/storage/traits/types.rs -> re-exports StorageEngineStrategy/StorageFormatStrategy

Note: src/storage/engine_capabilities.rs defines a SEPARATE EngineCapabilities
unit struct (a static-method holder that consumes CapabilityFactory) — it is
NOT the trait, is untouched, and still resolves via the shim.

Gates: cargo check --lib --bins + --tests -p proximadb clean; clippy --lib --bins
exit 0; fmt --check exit 0; 13/13 storage-ports tests pass; 20/20 consumer
capability lib tests pass via the shim.
