+++ crates/storage/proximadb-storage-ports/Cargo.toml
@@ -16,6 +16,7 @@ path = "src/lib.rs"
 [dependencies]
 anyhow.workspace = true
 async-trait.workspace = true
+serde.workspace = true
 proximadb-proto.workspace = true
 proximadb-storage-common.workspace = true
 # Foundation-tier: the StorageQuantizationEnginePort returns these data types
